backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.0
Flask-Limiter==3.5.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

